pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codebase-analyst"
version = "1.1.0"
description = "AI-powered codebase analysis and documentation generator"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["codebase", "analysis", "documentation", "ai", "langchain"]
authors = [
    {name = "Codebase Analyst Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Code Generators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "langfuse>=2.0.0",
]

[project.scripts]
codebase-analyst = "src.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/codebase-analyst"
Documentation = "https://github.com/yourusername/codebase-analyst#readme"
Repository = "https://github.com/yourusername/codebase-analyst"
Issues = "https://github.com/yourusername/codebase-analyst/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]
